Fix ndn-face-l2 clippy lints: unnecessary casts and dead_code
Remove redundant `as u16` casts on NDN_ETHERTYPE (already u16) in
af_packet, ether, and multicast_ether. Allow dead_code on
EtherMedium::local_mac which is stored for future use.

+++ crates/ndn-face-l2/src/af_packet.rs
@@ -57,7 +57,7 @@ pub fn open_packet_socket(ifindex: i32, protocol: u16) -> std::io::Result<OwnedF
         libc::socket(
             libc::AF_PACKET,
             libc::SOCK_DGRAM | libc::SOCK_NONBLOCK | libc::SOCK_CLOEXEC,
-            (protocol as u16).to_be() as i32,
+            protocol.to_be() as i32,
         )
     };
     if fd == -1 {
